// src/app/interfaces/common/our-vision.interface.ts
export interface OurVision {
  _id?: string;
  name?: string;
  image?: string;
  seoImage?: string;
  images?: string[];
  informations?: any[];
  description?: string;
  isHtml?: boolean;
  isHtmlEn?: boolean;
  websiteUrl?: string;
  slug?: string;
  createdAt?: Date;
  updatedAt?: Date;


  nameEn?: string;
 
  descriptionEn?: string;
}